requirements.txt
pandas
openpyxl
sentence-transformers
faiss-cpu
gradio
numpy